New shared value for cargo-packager identifier, single-instance unique id, and eframe app_id. Add a constant for the app name.
These should all be a unique id for the application, so we're using `com.mountainlizard.mountain_tiles` for all of them - basically the crate name with underscore rather than dash, plus domain name. We use the hyphen because this is how rust processes the crate name, and also because the freedesktop.org docs for the app id used in desktop files recommend not using dashes since they aren't compatible with dbus or flatpak.