Prepare for fg+bg tile colors
Add `TileColor` variant for fg+bg colors using palette indices
Rename functions for converting `TileColor` to a single color, to indicate that they operate on the foreground color (previously this was all there was to use, but now we may have background).
Increment `FileFormat` to v1 (only change is the new `TileColor` variant), and support loading v0 and v1 data by just treating it all as v1 (v0 contents are a subset of v1).